benchmark/requirements.txt
# MRAG-Bench evaluation dependencies
datasets>=2.14.0
gradio_client>=0.6.0
tqdm>=4.65.0
pandas>=2.0.0
Pillow>=9.0.0
